Scaffold vtcode-exec-events crate
