+++ docs/component_extraction_todo.md
@@ -52,7 +52,7 @@ This list tracks actionable tasks spawned from the component extraction plan as 
 
 ## `vtcode-exec-events`
 - [x] Outline extraction strategy covering event schema versioning, telemetry integration points, and feature gating (see `docs/vtcode_exec_events_extraction.md`).
-- [ ] Scaffold the standalone crate with the existing event enums and serialization helpers.
+- [x] Scaffold the standalone crate with the existing event enums and serialization helpers.
 - [ ] Document event semantics, versioning policy, and consumer integration patterns.
 - [ ] Publish examples or tests that demonstrate emitting and capturing execution events outside VTCode's runtime.
