+++ vtcode-core/src/exec/events.rs
@@ -1,187 +1,3 @@
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
+//! Re-exports structured execution events from the standalone crate.
 
-/// Structured events emitted during autonomous execution.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(tag = "type")]
-pub enum ThreadEvent {
-    /// Indicates that a new execution thread has started.
-    #[serde(rename = "thread.started")]
-    ThreadStarted(ThreadStartedEvent),
-    /// Marks the beginning of an execution turn.
-    #[serde(rename = "turn.started")]
-    TurnStarted(TurnStartedEvent),
-    /// Marks the completion of an execution turn.
-    #[serde(rename = "turn.completed")]
-    TurnCompleted(TurnCompletedEvent),
-    /// Marks a turn as failed with additional context.
-    #[serde(rename = "turn.failed")]
-    TurnFailed(TurnFailedEvent),
-    /// Indicates that an item has started processing.
-    #[serde(rename = "item.started")]
-    ItemStarted(ItemStartedEvent),
-    /// Indicates that an item has been updated.
-    #[serde(rename = "item.updated")]
-    ItemUpdated(ItemUpdatedEvent),
-    /// Indicates that an item reached a terminal state.
-    #[serde(rename = "item.completed")]
-    ItemCompleted(ItemCompletedEvent),
-    /// Represents a fatal error.
-    #[serde(rename = "error")]
-    Error(ThreadErrorEvent),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ThreadStartedEvent {
-    pub thread_id: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct TurnStartedEvent {}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct TurnCompletedEvent {
-    pub usage: Usage,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct TurnFailedEvent {
-    pub message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub usage: Option<Usage>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ThreadErrorEvent {
-    pub message: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-pub struct Usage {
-    pub input_tokens: u64,
-    pub cached_input_tokens: u64,
-    pub output_tokens: u64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ItemCompletedEvent {
-    pub item: ThreadItem,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ItemStartedEvent {
-    pub item: ThreadItem,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ItemUpdatedEvent {
-    pub item: ThreadItem,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ThreadItem {
-    pub id: String,
-    #[serde(flatten)]
-    pub details: ThreadItemDetails,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum ThreadItemDetails {
-    AgentMessage(AgentMessageItem),
-    Reasoning(ReasoningItem),
-    CommandExecution(CommandExecutionItem),
-    FileChange(FileChangeItem),
-    McpToolCall(McpToolCallItem),
-    WebSearch(WebSearchItem),
-    Error(ErrorItem),
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct AgentMessageItem {
-    pub text: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ReasoningItem {
-    pub text: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum CommandExecutionStatus {
-    #[default]
-    Completed,
-    Failed,
-    InProgress,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct CommandExecutionItem {
-    pub command: String,
-    #[serde(default)]
-    pub aggregated_output: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub exit_code: Option<i32>,
-    pub status: CommandExecutionStatus,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FileChangeItem {
-    pub changes: Vec<FileUpdateChange>,
-    pub status: PatchApplyStatus,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FileUpdateChange {
-    pub path: String,
-    pub kind: PatchChangeKind,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum PatchApplyStatus {
-    Completed,
-    Failed,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum PatchChangeKind {
-    Add,
-    Delete,
-    Update,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct McpToolCallItem {
-    pub tool_name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub arguments: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<McpToolCallStatus>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum McpToolCallStatus {
-    Started,
-    Completed,
-    Failed,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct WebSearchItem {
-    pub query: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub provider: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub results: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ErrorItem {
-    pub message: String,
-}
+pub use vtcode_exec_events::*;
