+++ vtcode-exec-events/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "vtcode-exec-events"
+version = "0.30.0"
+edition = "2024"
+description = "Structured execution telemetry event schema used across VTCode crates."
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/vinhnx/vtcode"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+[lints]
+workspace = true
