+++ vtcode-exec-events/src/lib.rs
@@ -1,0 +1,265 @@
+//! Structured execution telemetry events shared across VTCode crates.
+//!
+//! This crate exposes the serialized schema for thread lifecycle updates,
+//! command execution results, and other timeline artifacts emitted by the
+//! automation runtime. Downstream applications can deserialize these
+//! structures to drive dashboards, logging, or auditing pipelines without
+//! depending on the full `vtcode-core` crate.
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+/// Structured events emitted during autonomous execution.
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(tag = "type")]
+pub enum ThreadEvent {
+    /// Indicates that a new execution thread has started.
+    #[serde(rename = "thread.started")]
+    ThreadStarted(ThreadStartedEvent),
+    /// Marks the beginning of an execution turn.
+    #[serde(rename = "turn.started")]
+    TurnStarted(TurnStartedEvent),
+    /// Marks the completion of an execution turn.
+    #[serde(rename = "turn.completed")]
+    TurnCompleted(TurnCompletedEvent),
+    /// Marks a turn as failed with additional context.
+    #[serde(rename = "turn.failed")]
+    TurnFailed(TurnFailedEvent),
+    /// Indicates that an item has started processing.
+    #[serde(rename = "item.started")]
+    ItemStarted(ItemStartedEvent),
+    /// Indicates that an item has been updated.
+    #[serde(rename = "item.updated")]
+    ItemUpdated(ItemUpdatedEvent),
+    /// Indicates that an item reached a terminal state.
+    #[serde(rename = "item.completed")]
+    ItemCompleted(ItemCompletedEvent),
+    /// Represents a fatal error.
+    #[serde(rename = "error")]
+    Error(ThreadErrorEvent),
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ThreadStartedEvent {
+    /// Unique identifier for the thread that was started.
+    pub thread_id: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
+pub struct TurnStartedEvent {}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct TurnCompletedEvent {
+    /// Token usage summary for the completed turn.
+    pub usage: Usage,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct TurnFailedEvent {
+    /// Human-readable explanation describing why the turn failed.
+    pub message: String,
+    /// Optional token usage that was consumed before the failure occurred.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub usage: Option<Usage>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ThreadErrorEvent {
+    /// Fatal error message associated with the thread.
+    pub message: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
+pub struct Usage {
+    /// Number of prompt tokens processed during the turn.
+    pub input_tokens: u64,
+    /// Number of cached prompt tokens reused from previous turns.
+    pub cached_input_tokens: u64,
+    /// Number of completion tokens generated by the model.
+    pub output_tokens: u64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ItemCompletedEvent {
+    /// Snapshot of the thread item that completed.
+    pub item: ThreadItem,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ItemStartedEvent {
+    /// Snapshot of the thread item that began processing.
+    pub item: ThreadItem,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ItemUpdatedEvent {
+    /// Snapshot of the thread item after it was updated.
+    pub item: ThreadItem,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ThreadItem {
+    /// Stable identifier associated with the item.
+    pub id: String,
+    /// Embedded event details for the item type.
+    #[serde(flatten)]
+    pub details: ThreadItemDetails,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(tag = "type", rename_all = "snake_case")]
+pub enum ThreadItemDetails {
+    /// Message authored by the agent.
+    AgentMessage(AgentMessageItem),
+    /// Free-form reasoning text produced during a turn.
+    Reasoning(ReasoningItem),
+    /// Command execution lifecycle update.
+    CommandExecution(CommandExecutionItem),
+    /// File change summary associated with the turn.
+    FileChange(FileChangeItem),
+    /// MCP tool invocation status.
+    McpToolCall(McpToolCallItem),
+    /// Web search event emitted by a registered search provider.
+    WebSearch(WebSearchItem),
+    /// General error captured for auditing.
+    Error(ErrorItem),
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct AgentMessageItem {
+    /// Textual content of the agent message.
+    pub text: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ReasoningItem {
+    /// Free-form reasoning content captured during planning.
+    pub text: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
+#[serde(rename_all = "snake_case")]
+pub enum CommandExecutionStatus {
+    /// Command finished successfully.
+    #[default]
+    Completed,
+    /// Command failed (non-zero exit code or runtime error).
+    Failed,
+    /// Command is still running and may emit additional output.
+    InProgress,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct CommandExecutionItem {
+    /// Command string executed by the runner.
+    pub command: String,
+    /// Aggregated output emitted by the command.
+    #[serde(default)]
+    pub aggregated_output: String,
+    /// Exit code reported by the process, when available.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub exit_code: Option<i32>,
+    /// Current status of the command execution.
+    pub status: CommandExecutionStatus,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct FileChangeItem {
+    /// List of individual file updates included in the change set.
+    pub changes: Vec<FileUpdateChange>,
+    /// Whether the patch application succeeded.
+    pub status: PatchApplyStatus,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct FileUpdateChange {
+    /// Path of the file that was updated.
+    pub path: String,
+    /// Type of change applied to the file.
+    pub kind: PatchChangeKind,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum PatchApplyStatus {
+    /// Patch successfully applied.
+    Completed,
+    /// Patch application failed.
+    Failed,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum PatchChangeKind {
+    /// File addition.
+    Add,
+    /// File deletion.
+    Delete,
+    /// File update in place.
+    Update,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct McpToolCallItem {
+    /// Name of the MCP tool invoked by the agent.
+    pub tool_name: String,
+    /// Arguments passed to the tool invocation, if any.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub arguments: Option<Value>,
+    /// Result payload returned by the tool, if captured.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub result: Option<String>,
+    /// Lifecycle status for the tool call.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub status: Option<McpToolCallStatus>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum McpToolCallStatus {
+    /// Tool invocation has started.
+    Started,
+    /// Tool invocation completed successfully.
+    Completed,
+    /// Tool invocation failed.
+    Failed,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct WebSearchItem {
+    /// Query that triggered the search.
+    pub query: String,
+    /// Search provider identifier, when known.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub provider: Option<String>,
+    /// Optional raw search results captured for auditing.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub results: Option<Vec<String>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct ErrorItem {
+    /// Error message displayed to the user or logs.
+    pub message: String,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn thread_event_round_trip() {
+        let event = ThreadEvent::TurnCompleted(TurnCompletedEvent {
+            usage: Usage {
+                input_tokens: 1,
+                cached_input_tokens: 2,
+                output_tokens: 3,
+            },
+        });
+
+        let json = serde_json::to_string(&event).expect("serialize");
+        let restored: ThreadEvent = serde_json::from_str(&json).expect("deserialize");
+
+        assert_eq!(restored, event);
+    }
+}
